Move GidObject to separate file/header
