+++ email/include/email/gid.hpp
@@ -99,38 +99,9 @@ private:
   const std::string value_string_;
 };
 
-/// Abstract object with a GID.
-/**
- * TODO(christophebedard) move to its own header
- */
-class GidObject
-{
-public:
-  /// Get the unique ID (GID) for this object.
-  /**
-   * \return the GID
-   */
-  EMAIL_PUBLIC
-  const Gid &
-  get_gid() const;
-
-protected:
-  /// Constructor
-  EMAIL_PUBLIC
-  GidObject();
-
-  EMAIL_PUBLIC
-  virtual ~GidObject();
-
-private:
-  EMAIL_DISABLE_COPY(GidObject)
-
-  const Gid gid_;
-};
-
 }  // namespace email
 
-/// Formatting for Gid objects.
+/// Formatting for GIDs.
 template<>
 struct fmt::formatter<email::Gid>: formatter<string_view>
 {
