+++ email/src/service.cpp
@@ -16,6 +16,7 @@
 #include <string>
 
 #include "email/gid.hpp"
+#include "email/gid_object.hpp"
 #include "email/named_object.hpp"
 #include "email/service.hpp"
